+++ examples/virtualbackground/index.js
@@ -15,6 +15,9 @@ const defaultParams = {
   capResolution: '1280x720',
   deferInputFrameDownscale: 'false',
   maskBlurRadius: '8',
+  hysteresisEnabled: 'true',
+  hysteresisHighThreshold: '180',
+  hysteresisLowThreshold: '80',
   stats: 'show',
   useWebWorker: 'true',
 };
@@ -60,6 +63,9 @@ const loadImage = async (name) => {
     capFramerate,
     capResolution,
     deferInputFrameDownscale,
+    hysteresisEnabled,
+    hysteresisHighThreshold,
+    hysteresisLowThreshold,
     maskBlurRadius,
     stats,
     useWebWorker,
@@ -83,6 +89,9 @@ const loadImage = async (name) => {
   const processorOptions = {
     assetsPath,
     deferInputFrameDownscale: JSON.parse(deferInputFrameDownscale),
+    hysteresisEnabled: JSON.parse(hysteresisEnabled),
+    hysteresisHighThreshold: Number(hysteresisHighThreshold),
+    hysteresisLowThreshold: Number(hysteresisLowThreshold),
     maskBlurRadius: Number(maskBlurRadius),
     useWebWorker: JSON.parse(useWebWorker),
   };