+++ lib/processors/background/VirtualBackgroundProcessor.ts
@@ -1,5 +1,5 @@
 import { ImageFit } from '../../types';
-import { MASK_BLUR_RADIUS } from '../../constants';
+import { HYSTERESIS_ENABLED, HYSTERESIS_HIGH, HYSTERESIS_LOW, MASK_BLUR_RADIUS } from '../../constants';
 import { BackgroundProcessor, BackgroundProcessorOptions } from './BackgroundProcessor';
 import { VirtualBackgroundProcessorPipeline, VirtualBackgroundProcessorPipelineProxy } from './pipelines/backgroundprocessorpipeline';
 
@@ -86,6 +86,9 @@ export class VirtualBackgroundProcessor extends BackgroundProcessor<VirtualBackg
       backgroundImage,
       deferInputFrameDownscale = false,
       fitType = ImageFit.Fill,
+      hysteresisEnabled = HYSTERESIS_ENABLED,
+      hysteresisHighThreshold = HYSTERESIS_HIGH,
+      hysteresisLowThreshold = HYSTERESIS_LOW,
       maskBlurRadius = MASK_BLUR_RADIUS,
       useWebWorker = true
     } = options;
@@ -103,6 +106,9 @@ export class VirtualBackgroundProcessor extends BackgroundProcessor<VirtualBackg
       assetsPath,
       deferInputFrameDownscale,
       fitType,
+      hysteresisEnabled,
+      hysteresisHighThreshold,
+      hysteresisLowThreshold,
       maskBlurRadius
     });